Refactor (package): rename SDK package to avoid naming conflicts
