+++ python/examples/nodepools_config.py
