+++ python/rackspace_spot_sdk/classes.py
